1.1.1/requirements.txt
requests
beautifulsoup4
pillow
pytesseract